Create Accretion Disk Modelling
This function was used to determine a chi value for the quality of the fit of the model to the data. In this case, we were modelling the example spectra of an accretion disc, a White Dwarf and a small M-type star to a 3 Band photometric observation on the Cataclysmic variable star DW Cancri.